requirements.txt
asyncpg==0.27.0
greenlet==2.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
